chore(RHINENG-15653): Cleanup CompliancePolicies
